+++ src/main/java/ru/hse/plugin/metrics/git/SpecificBranchCommitCounter.java
@@ -6,7 +6,7 @@ import ru.hse.plugin.metrics.abstracts.Metric;
 
 import java.util.Objects;
 
-import static ru.hse.plugin.metrics.commons.Names.SPECIFIC_BRANCH_COMMIT_COUNTER;
+import static ru.hse.plugin.metrics.commons.util.Names.SPECIFIC_BRANCH_COMMIT_COUNTER;
 
 public class SpecificBranchCommitCounter extends CountingMetric {
     @NotNull
